Module embedded_tls::blocking

Re-exports

pub use crate::TlsError;

Structs

Aes128GcmSha256

Aes256GcmSha384

NoClock

NoVerify

SplitConnectionState

TlsConfig

TlsConnection

Type representing a TLS connection. An instance of this type can be used to establish a TLS connection, write and read encrypted data over this connection, and closing to free up the underlying resources.

TlsContext

TlsReader

TlsWriter

Enums

Certificate

MaxFragmentLength

Maximum plaintext fragment length

Constants

TLS_RECORD_OVERHEAD

Traits

TlsCipherSuite

Represents a TLS 1.3 cipher suite

TlsClock

TlsVerifier

A TLS 1.3 verifier.

Type Aliases

Sha256

SHA-256 hasher.

Sha384

SHA-384 hasher.